moved changes to separate service, to prevent changes in all AssetRatioService relates modules
