+++ browser/ui/webui/brave_wallet/wallet_page_ui.cc
@@ -14,6 +14,7 @@
 #include "brave/browser/brave_wallet/bitcoin_wallet_service_factory.h"
 #include "brave/browser/brave_wallet/brave_wallet_ipfs_service_factory.h"
 #include "brave/browser/brave_wallet/brave_wallet_service_factory.h"
+#include "brave/browser/brave_wallet/buy_and_sell_service_factory.h"
 #include "brave/browser/brave_wallet/json_rpc_service_factory.h"
 #include "brave/browser/brave_wallet/keyring_service_factory.h"
 #include "brave/browser/brave_wallet/swap_service_factory.h"
@@ -135,7 +136,9 @@ void WalletPageUI::CreatePageHandler(
     mojo::PendingReceiver<brave_wallet::mojom::WalletAutoPinService>
         brave_wallet_auto_pin_service_receiver,
     mojo::PendingReceiver<brave_wallet::mojom::IpfsService>
-        ipfs_service_receiver) {
+        ipfs_service_receiver,
+    mojo::PendingReceiver<brave_wallet::mojom::BuyAndSellService>
+        buy_and_sell_service) {
   DCHECK(page);
   auto* profile = Profile::FromWebUI(web_ui());
   DCHECK(profile);
@@ -155,6 +158,8 @@ void WalletPageUI::CreatePageHandler(
       profile, std::move(swap_service_receiver));
   brave_wallet::AssetRatioServiceFactory::BindForContext(
       profile, std::move(asset_ratio_service_receiver));
+  brave_wallet::BuyAndSellServiceFactory::BindForContext(
+      profile, std::move(buy_and_sell_service));
   brave_wallet::KeyringServiceFactory::BindForContext(
       profile, std::move(keyring_service_receiver));
   brave_wallet::TxServiceFactory::BindForContext(
