+++ browser/ui/webui/brave_wallet/wallet_page_ui.h
@@ -68,7 +68,9 @@ class WalletPageUI : public ui::MojoWebUIController,
       mojo::PendingReceiver<brave_wallet::mojom::WalletAutoPinService>
           brave_wallet_auto_pin_service_receiver,
       mojo::PendingReceiver<brave_wallet::mojom::IpfsService>
-          brave_wallet_ipfs_service_receiver) override;
+          brave_wallet_ipfs_service_receiver,
+      mojo::PendingReceiver<brave_wallet::mojom::BuyAndSellService>
+          buy_and_sell_service) override;
 
   std::unique_ptr<WalletPageHandler> page_handler_;
   std::unique_ptr<brave_wallet::WalletHandler> wallet_handler_;
