+++ components/brave_wallet/browser/BUILD.gn
@@ -75,6 +75,10 @@ static_library("browser") {
     "brave_wallet_service_delegate.cc",
     "brave_wallet_service_delegate.h",
     "brave_wallet_service_observer_base.h",
+    "buy_and_sell_response_parser.cc",
+    "buy_and_sell_response_parser.h",
+    "buy_and_sell_service.cc",
+    "buy_and_sell_service.h",
     "ens_resolver_task.cc",
     "ens_resolver_task.h",
     "eth_abi_decoder.cc",
