+++ components/brave_wallet/browser/asset_ratio_service.cc
@@ -127,21 +127,6 @@ base::flat_map<std::string, std::string> MakeBraveServicesKeyHeader() {
   return request_headers;
 }
 
-base::flat_map<std::string, std::string> MakeMeldApiHeaders() {
-  base::flat_map<std::string, std::string> request_headers;
-  std::unique_ptr<base::Environment> env(base::Environment::Create());
-  std::string meld_api_key(BUILDFLAG(MELD_API_KEY));
-  if (env->HasVar("MELD_API_KEY")) {
-    env->GetVar("MELD_API_KEY", &meld_api_key);
-  }
-  meld_api_key = base::StrCat({"BASIC", " ", meld_api_key});
-  request_headers["Authorization"] = std::move(meld_api_key);
-  request_headers["accept"] = "application/json";
-
-  return request_headers;
-}
-constexpr char kDefaultMeldStatuses[] = "LIVE,RECENTLY_ADDED";
-
 }  // namespace
 
 namespace brave_wallet {
@@ -395,7 +380,6 @@ void AssetRatioService::GetPrice(const std::vector<std::string>& from_assets,
                                  const std::vector<std::string>& to_assets,
                                  mojom::AssetPriceTimeframe timeframe,
                                  GetPriceCallback callback) {
-  MakeMeldApiHeaders();
   if (dummy_prices_for_testing_) {
     std::move(callback).Run(true, DummyPrices(from_assets, to_assets));
     return;
@@ -578,490 +562,6 @@ void AssetRatioService::OnGetCoinMarkets(GetCoinMarketsCallback callback,
   }
 
   std::move(callback).Run(true, std::move(*values));
-}
-
-// static
-GURL AssetRatioService::GetServiceProviderURL(
-    const std::string& countries,
-    const std::string& fiat_currencies,
-    const std::string& crypto_currencies,
-    const std::string& service_providers,
-    const std::string& payment_method_types,
-    const std::string& statuses) {
-  auto url = GURL(base::StringPrintf("%s/service-providers",
-                                     base_url_for_test_.is_empty()
-                                         ? GetMeldAssetRatioBaseURL().c_str()
-                                         : base_url_for_test_.spec().c_str()));
-  url = net::AppendQueryParameter(url, "accountFilter", "false");
-  if (!statuses.empty()) {
-    url = net::AppendQueryParameter(url, "statuses", statuses);
-  } else {
-    url = net::AppendQueryParameter(url, "statuses", kDefaultMeldStatuses);
-  }
-
-  if (!countries.empty()) {
-    url = net::AppendQueryParameter(url, "countries", countries);
-  }
-  if (!fiat_currencies.empty()) {
-    url = net::AppendQueryParameter(url, "fiatCurrencies", fiat_currencies);
-  }
-  if (!crypto_currencies.empty()) {
-    url = net::AppendQueryParameter(url, "cryptoCurrencies", crypto_currencies);
-  }
-  if (!service_providers.empty()) {
-    url = net::AppendQueryParameter(url, "serviceProviders", service_providers);
-  }
-  if (!payment_method_types.empty()) {
-    url = net::AppendQueryParameter(url, "paymentMethodTypes",
-                                    payment_method_types);
-  }
-
-  return url;
-}
-
-void AssetRatioService::GetServiceProviders(
-    const std::string& countries,
-    const std::string& fiat_currencies,
-    const std::string& crypto_currencies,
-    const std::string& service_providers,
-    const std::string& payment_method_types,
-    const std::string& statuses,
-    GetServiceProvidersCallback callback) {
-  auto internal_callback =
-      base::BindOnce(&AssetRatioService::OnGetServiceProviders,
-                     weak_ptr_factory_.GetWeakPtr(), std::move(callback));
-
-  api_request_helper_->Request(
-      "GET",
-      GetServiceProviderURL(countries, fiat_currencies, crypto_currencies,
-                            service_providers, payment_method_types, statuses),
-      "", "", std::move(internal_callback), MakeMeldApiHeaders(),
-      {.auto_retry_on_network_change = true, .enable_cache = true});
-}
-
-void AssetRatioService::OnGetServiceProviders(
-    GetServiceProvidersCallback callback,
-    APIRequestResult api_request_result) {
-  if (!api_request_result.Is2XXResponseCode()) {
-    std::move(callback).Run({},
-                            std::vector<std::string>{"INTERNAL_SERVICE_ERROR"});
-    return;
-  }
-
-  if (std::vector<std::string> errors;
-      ParseMeldErrorResponse(api_request_result.value_body(), &errors)) {
-    std::move(callback).Run({}, errors);
-    return;
-  }
-
-  std::vector<mojom::ServiceProviderPtr> service_providers;
-  if (!ParseServiceProviders(api_request_result.value_body(),
-                             &service_providers)) {
-    std::move(callback).Run({}, std::vector<std::string>{"PARSING_ERROR"});
-    return;
-  }
-  std::move(callback).Run(std::move(service_providers), std::nullopt);
-}
-
-void AssetRatioService::GetCryptoQuotes(
-    const std::string& country,
-    const std::string& source_currency_code,
-    const std::string& destination_currency_code,
-    const double source_amount,
-    const std::string& account,
-    GetCryptoQuotesCallback callback) {
-  base::Value::Dict payload;
-  AddKeyIfNotEmpty(&payload, "countryCode", country);
-  AddKeyIfNotEmpty(&payload, "sourceCurrencyCode", source_currency_code);
-  AddKeyIfNotEmpty(&payload, "sourceAmount",
-                   base::NumberToString(source_amount));
-  AddKeyIfNotEmpty(&payload, "destinationCurrencyCode",
-                   destination_currency_code);
-  AddKeyIfNotEmpty(&payload, "walletAddress", account);
-
-  const std::string json_payload = GetJSON(payload);
-
-  auto url = GURL(base::StringPrintf("%s/payments/crypto/quote",
-                                     base_url_for_test_.is_empty()
-                                         ? GetMeldAssetRatioBaseURL().c_str()
-                                         : base_url_for_test_.spec().c_str()));
-  auto internal_callback =
-      base::BindOnce(&AssetRatioService::OnGetCryptoQuotes,
-                     weak_ptr_factory_.GetWeakPtr(), std::move(callback));
-
-  api_request_helper_->Request(
-      "POST", url, json_payload, "application/json",
-      std::move(internal_callback), MakeMeldApiHeaders(),
-      {.auto_retry_on_network_change = true, .enable_cache = false});
-}
-
-void AssetRatioService::OnGetCryptoQuotes(GetCryptoQuotesCallback callback,
-                                          APIRequestResult api_request_result) {
-  if (!api_request_result.Is2XXResponseCode()) {
-    std::move(callback).Run({},
-                            std::vector<std::string>{"INTERNAL_SERVICE_ERROR"});
-    return;
-  }
-
-  if (std::vector<std::string> errors;
-      ParseMeldErrorResponse(api_request_result.value_body(), &errors)) {
-    std::move(callback).Run({}, errors);
-    return;
-  }
-
-  std::optional<std::vector<std::string>> errors;
-  std::string error;
-  std::vector<mojom::CryptoQuotePtr> quotes;
-  if (!ParseCryptoQuotes(api_request_result.value_body(), &quotes, &error)) {
-    errors = std::vector<std::string>{"PARSING_ERROR"};
-    std::move(callback).Run({}, errors);
-    return;
-  }
-
-  if (!error.empty()) {
-    errors = std::vector<std::string>{error};
-  }
-
-  std::move(callback).Run(std::move(quotes), errors);
-}
-
-// static
-GURL AssetRatioService::GetGetPaymentMethodsURL(
-    const std::string& countries,
-    const std::string& fiat_currencies,
-    const std::string& crypto_currencies,
-    const std::string& service_providers,
-    const std::string& payment_method_types,
-    const std::string& statuses) {
-  auto url = GURL(base::StringPrintf(
-      "%s/service-providers/properties/payment-methods",
-      base_url_for_test_.is_empty() ? GetMeldAssetRatioBaseURL().c_str()
-                                    : base_url_for_test_.spec().c_str()));
-  url =
-      net::AppendQueryParameter(url, "includeServiceProviderDetails", "false");
-  if (!statuses.empty()) {
-    url = net::AppendQueryParameter(url, "statuses", statuses);
-  } else {
-    url = net::AppendQueryParameter(url, "statuses", kDefaultMeldStatuses);
-  }
-
-  if (!countries.empty()) {
-    url = net::AppendQueryParameter(url, "countries", countries);
-  }
-  if (!fiat_currencies.empty()) {
-    url = net::AppendQueryParameter(url, "fiatCurrencies", fiat_currencies);
-  }
-  if (!crypto_currencies.empty()) {
-    url = net::AppendQueryParameter(url, "cryptoCurrencies", crypto_currencies);
-  }
-  if (!service_providers.empty()) {
-    url = net::AppendQueryParameter(url, "serviceProviders", service_providers);
-  }
-  if (!payment_method_types.empty()) {
-    url = net::AppendQueryParameter(url, "paymentMethodTypes",
-                                    payment_method_types);
-  }
-
-  return url;
-}
-
-void AssetRatioService::GetPaymentMethods(
-    const std::string& countries,
-    const std::string& fiat_currencies,
-    const std::string& crypto_currencies,
-    const std::string& service_providers,
-    const std::string& payment_method_types,
-    const std::string& statuses,
-    GetPaymentMethodsCallback callback) {
-  auto internal_callback =
-      base::BindOnce(&AssetRatioService::OnGetPaymentMethods,
-                     weak_ptr_factory_.GetWeakPtr(), std::move(callback));
-
-  api_request_helper_->Request(
-      "GET",
-      GetGetPaymentMethodsURL(countries, fiat_currencies, crypto_currencies,
-                              service_providers, payment_method_types,
-                              statuses),
-      "", "", std::move(internal_callback), MakeMeldApiHeaders(),
-      {.auto_retry_on_network_change = true, .enable_cache = true});
-}
-
-void AssetRatioService::OnGetPaymentMethods(
-    GetPaymentMethodsCallback callback,
-    APIRequestResult api_request_result) {
-  if (!api_request_result.Is2XXResponseCode()) {
-    std::move(callback).Run({},
-                            std::vector<std::string>{"INTERNAL_SERVICE_ERROR"});
-    return;
-  }
-
-  if (std::vector<std::string> errors;
-      ParseMeldErrorResponse(api_request_result.value_body(), &errors)) {
-    std::move(callback).Run({}, errors);
-    return;
-  }
-
-  std::vector<mojom::PaymentMethodPtr> payment_methods;
-  if (!ParsePaymentMethods(api_request_result.value_body(), &payment_methods)) {
-    std::move(callback).Run({}, std::vector<std::string>{"PARSING_ERROR"});
-    return;
-  }
-
-  std::move(callback).Run(std::move(payment_methods), std::nullopt);
-}
-
-// static
-GURL AssetRatioService::GetFiatCurrenciesURL(
-    const std::string& countries,
-    const std::string& fiat_currencies,
-    const std::string& crypto_currencies,
-    const std::string& service_providers,
-    const std::string& payment_method_types,
-    const std::string& statuses) {
-  auto url = GURL(base::StringPrintf(
-      "%s/service-providers/properties/fiat-currencies",
-      base_url_for_test_.is_empty() ? GetMeldAssetRatioBaseURL().c_str()
-                                    : base_url_for_test_.spec().c_str()));
-  url =
-      net::AppendQueryParameter(url, "includeServiceProviderDetails", "false");
-  if (!statuses.empty()) {
-    url = net::AppendQueryParameter(url, "statuses", statuses);
-  } else {
-    url = net::AppendQueryParameter(url, "statuses", kDefaultMeldStatuses);
-  }
-
-  if (!countries.empty()) {
-    url = net::AppendQueryParameter(url, "countries", countries);
-  }
-  if (!fiat_currencies.empty()) {
-    url = net::AppendQueryParameter(url, "fiatCurrencies", fiat_currencies);
-  }
-  if (!crypto_currencies.empty()) {
-    url = net::AppendQueryParameter(url, "cryptoCurrencies", crypto_currencies);
-  }
-  if (!service_providers.empty()) {
-    url = net::AppendQueryParameter(url, "serviceProviders", service_providers);
-  }
-  if (!payment_method_types.empty()) {
-    url = net::AppendQueryParameter(url, "paymentMethodTypes",
-                                    payment_method_types);
-  }
-
-  return url;
-}
-
-void AssetRatioService::GetFiatCurrencies(
-    const std::string& countries,
-    const std::string& fiat_currencies,
-    const std::string& crypto_currencies,
-    const std::string& service_providers,
-    const std::string& payment_method_types,
-    const std::string& statuses,
-    GetFiatCurrenciesCallback callback) {
-  auto internal_callback =
-      base::BindOnce(&AssetRatioService::OnGetFiatCurrencies,
-                     weak_ptr_factory_.GetWeakPtr(), std::move(callback));
-
-  api_request_helper_->Request(
-      "GET",
-      GetFiatCurrenciesURL(countries, fiat_currencies, crypto_currencies,
-                           service_providers, payment_method_types, statuses),
-      "", "", std::move(internal_callback), MakeMeldApiHeaders(),
-      {.auto_retry_on_network_change = true, .enable_cache = true});
-}
-
-void AssetRatioService::OnGetFiatCurrencies(
-    GetFiatCurrenciesCallback callback,
-    APIRequestResult api_request_result) {
-  if (!api_request_result.Is2XXResponseCode()) {
-    std::move(callback).Run({},
-                            std::vector<std::string>{"INTERNAL_SERVICE_ERROR"});
-    return;
-  }
-
-  if (std::vector<std::string> errors;
-      ParseMeldErrorResponse(api_request_result.value_body(), &errors)) {
-    std::move(callback).Run({}, errors);
-    return;
-  }
-
-  std::vector<mojom::FiatCurrencyPtr> fiat_currencies;
-  if (!ParseFiatCurrencies(api_request_result.value_body(), &fiat_currencies)) {
-    std::move(callback).Run({}, std::vector<std::string>{"PARSING_ERROR"});
-    return;
-  }
-
-  std::move(callback).Run(std::move(fiat_currencies), std::nullopt);
-}
-
-// static
-GURL AssetRatioService::GetCryptoCurrenciesURL(
-    const std::string& countries,
-    const std::string& fiat_currencies,
-    const std::string& crypto_currencies,
-    const std::string& service_providers,
-    const std::string& payment_method_types,
-    const std::string& statuses) {
-  auto url = GURL(base::StringPrintf(
-      "%s/service-providers/properties/crypto-currencies",
-      base_url_for_test_.is_empty() ? GetMeldAssetRatioBaseURL().c_str()
-                                    : base_url_for_test_.spec().c_str()));
-
-  url =
-      net::AppendQueryParameter(url, "includeServiceProviderDetails", "false");
-
-  if (!statuses.empty()) {
-    url = net::AppendQueryParameter(url, "statuses", statuses);
-  } else {
-    url = net::AppendQueryParameter(url, "statuses", kDefaultMeldStatuses);
-  }
-
-  if (!countries.empty()) {
-    url = net::AppendQueryParameter(url, "countries", countries);
-  }
-  if (!fiat_currencies.empty()) {
-    url = net::AppendQueryParameter(url, "fiatCurrencies", fiat_currencies);
-  }
-  if (!crypto_currencies.empty()) {
-    url = net::AppendQueryParameter(url, "cryptoCurrencies", crypto_currencies);
-  }
-  if (!service_providers.empty()) {
-    url = net::AppendQueryParameter(url, "serviceProviders", service_providers);
-  }
-  if (!payment_method_types.empty()) {
-    url = net::AppendQueryParameter(url, "paymentMethodTypes",
-                                    payment_method_types);
-  }
-
-  return url;
-}
-
-void AssetRatioService::GetCryptoCurrencies(
-    const std::string& countries,
-    const std::string& fiat_currencies,
-    const std::string& crypto_currencies,
-    const std::string& service_providers,
-    const std::string& payment_method_types,
-    const std::string& statuses,
-    GetCryptoCurrenciesCallback callback) {
-  auto internal_callback =
-      base::BindOnce(&AssetRatioService::OnGetCryptoCurrencies,
-                     weak_ptr_factory_.GetWeakPtr(), std::move(callback));
-
-  api_request_helper_->Request(
-      "GET",
-      GetCryptoCurrenciesURL(countries, fiat_currencies, crypto_currencies,
-                             service_providers, payment_method_types, statuses),
-      "", "", std::move(internal_callback), MakeMeldApiHeaders(),
-      {.auto_retry_on_network_change = true, .enable_cache = true});
-}
-
-void AssetRatioService::OnGetCryptoCurrencies(
-    GetCryptoCurrenciesCallback callback,
-    APIRequestResult api_request_result) {
-  if (!api_request_result.Is2XXResponseCode()) {
-    std::move(callback).Run({},
-                            std::vector<std::string>{"INTERNAL_SERVICE_ERROR"});
-    return;
-  }
-
-  if (std::vector<std::string> errors;
-      ParseMeldErrorResponse(api_request_result.value_body(), &errors)) {
-    std::move(callback).Run({}, errors);
-    return;
-  }
-
-  std::vector<mojom::CryptoCurrencyPtr> crypto_currencies;
-  if (!ParseCryptoCurrencies(api_request_result.value_body(),
-                             &crypto_currencies)) {
-    std::move(callback).Run({}, std::vector<std::string>{"PARSING_ERROR"});
-    return;
-  }
-
-  std::move(callback).Run(std::move(crypto_currencies), std::nullopt);
-}
-
-// static
-GURL AssetRatioService::GetCountriesURL(const std::string& countries,
-                                        const std::string& fiat_currencies,
-                                        const std::string& crypto_currencies,
-                                        const std::string& service_providers,
-                                        const std::string& payment_method_types,
-                                        const std::string& statuses) {
-  auto url = GURL(base::StringPrintf(
-      "%s/service-providers/properties/countries",
-      base_url_for_test_.is_empty() ? GetMeldAssetRatioBaseURL().c_str()
-                                    : base_url_for_test_.spec().c_str()));
-
-  url =
-      net::AppendQueryParameter(url, "includeServiceProviderDetails", "false");
-
-  if (!statuses.empty()) {
-    url = net::AppendQueryParameter(url, "statuses", statuses);
-  } else {
-    url = net::AppendQueryParameter(url, "statuses", kDefaultMeldStatuses);
-  }
-
-  if (!countries.empty()) {
-    url = net::AppendQueryParameter(url, "countries", countries);
-  }
-  if (!fiat_currencies.empty()) {
-    url = net::AppendQueryParameter(url, "fiatCurrencies", fiat_currencies);
-  }
-  if (!crypto_currencies.empty()) {
-    url = net::AppendQueryParameter(url, "cryptoCurrencies", crypto_currencies);
-  }
-  if (!service_providers.empty()) {
-    url = net::AppendQueryParameter(url, "serviceProviders", service_providers);
-  }
-  if (!payment_method_types.empty()) {
-    url = net::AppendQueryParameter(url, "paymentMethodTypes",
-                                    payment_method_types);
-  }
-
-  return url;
-}
-
-void AssetRatioService::GetCountries(const std::string& countries,
-                                     const std::string& fiat_currencies,
-                                     const std::string& crypto_currencies,
-                                     const std::string& service_providers,
-                                     const std::string& payment_method_types,
-                                     const std::string& statuses,
-                                     GetCountriesCallback callback) {
-  auto internal_callback =
-      base::BindOnce(&AssetRatioService::OnGetCountries,
-                     weak_ptr_factory_.GetWeakPtr(), std::move(callback));
-
-  api_request_helper_->Request(
-      "GET",
-      GetCountriesURL(countries, fiat_currencies, crypto_currencies,
-                      service_providers, payment_method_types, statuses),
-      "", "", std::move(internal_callback), MakeMeldApiHeaders(),
-      {.auto_retry_on_network_change = true, .enable_cache = true});
-}
-
-void AssetRatioService::OnGetCountries(GetCountriesCallback callback,
-                                       APIRequestResult api_request_result) {
-  if (!api_request_result.Is2XXResponseCode()) {
-    std::move(callback).Run({},
-                            std::vector<std::string>{"INTERNAL_SERVICE_ERROR"});
-    return;
-  }
-
-  if (std::vector<std::string> errors;
-      ParseMeldErrorResponse(api_request_result.value_body(), &errors)) {
-    std::move(callback).Run({}, errors);
-    return;
-  }
-
-  std::vector<mojom::CountryPtr> crypto_currencies;
-  if (!ParseCountries(api_request_result.value_body(), &crypto_currencies)) {
-    std::move(callback).Run({}, std::vector<std::string>{"PARSING_ERROR"});
-    return;
-  }
-
-  std::move(callback).Run(std::move(crypto_currencies), std::nullopt);
 }
 
 }  // namespace brave_wallet
