+++ components/brave_wallet/browser/test/BUILD.gn
@@ -30,6 +30,8 @@ source_set("brave_wallet_unit_tests") {
     "//brave/components/brave_wallet/browser/blockchain_registry_unittest.cc",
     "//brave/components/brave_wallet/browser/brave_wallet_p3a_unittest.cc",
     "//brave/components/brave_wallet/browser/brave_wallet_utils_unittest.cc",
+    "//brave/components/brave_wallet/browser/buy_and_sell_response_parser_unittest.cc",
+    "//brave/components/brave_wallet/browser/buy_and_sell_service_unittest.cc",
     "//brave/components/brave_wallet/browser/eip1559_transaction_unittest.cc",
     "//brave/components/brave_wallet/browser/eip2930_transaction_unittest.cc",
     "//brave/components/brave_wallet/browser/eth_abi_decoder_unittest.cc",
